Update install script msg

+++ install.sh
@@ -2,6 +2,7 @@
 set -eu
 
 GREEN='\033[0;32m'
+RED='\033[0;31m'
 NC='\033[0m' # No Color
 INSTALL_DIR="$(realpath "${1:-$HOME/.deven}")"
 
@@ -15,8 +16,14 @@ else
 fi
 
 local_bin="$HOME"/.local/bin
-printf "${GREEN}Softlinking deven binary to %s${NC}\n" "$local_bin"
-chmod +x "$INSTALL_DIR"/deven.sh
 mkdir -p "$local_bin"
-ln -s "$INSTALL_DIR"/deven.sh "$local_bin"/deven
-printf "${GREEN}deven installed successfully, you might need to add ~/.local/bin/ to PATH to use it${NC}\n"
+if [ ! -f "$local_bin"/deven ]; then
+    printf "${GREEN}Softlinking deven to %s${NC}\n" "$local_bin"
+    chmod +x "$INSTALL_DIR"/deven.sh
+    ln -s "$INSTALL_DIR"/deven.sh "$local_bin"/deven
+else
+    printf "${RED}ERROR: deven already exists in %s${NC}\n" "$local_bin"
+    failed=true
+fi
+
+[ "${failed-}" != "true" ] && printf "${GREEN}deven installed successfully, you might need to add ~/.local/bin/ to PATH to use it${NC}\n"
